+++ src/f_stdio/printf/cast.c
@@ -6,7 +6,7 @@
 /*   By: abarthel <abarthel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/05 14:54:14 by abarthel          #+#    #+#             */
-/*   Updated: 2019/04/18 12:16:45 by abarthel         ###   ########.fr       */
+/*   Updated: 2020/06/26 21:02:52 by abarthel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ void	ft_cast_unbr(uintmax_t *__restrict__ nb)
 	else if (g_modifier.l)
 		*nb = (unsigned long)(*nb);
 	else if (g_modifier.ll)
-		*nb = (unsigned long long)*nb;
+		*nb = (uint64_t)*nb;
 	else if (g_modifier.j)
 		*nb = (uintmax_t)(*nb);
 	else if (g_modifier.t)
@@ -47,7 +47,7 @@ void	ft_cast_nbr(intmax_t *__restrict__ nb)
 	else if (g_modifier.l)
 		*nb = (long)(*nb);
 	else if (g_modifier.ll)
-		*nb = (long long)*nb;
+		*nb = (int64_t)*nb;
 	else if (g_modifier.j)
 		*nb = (intmax_t)(*nb);
 	else if (g_modifier.t)
